Fit regression model under Null Hypothesis

NullModel fits a linear, generalized or mixed regression model with random effects specified by covariance matrices (See fitNullModel) such as a kinship matrix. The output of NullModel can be passed to ewas_meth_exposure for association testing.

NullModel(
  db_obj,
  trait,
  covariates_string = NULL,
  cov.mat = NULL,
  method = "lm",
  family = FALSE
)

Arguments

db_obj

MethList created using create_methlist.

trait

Outcome phenotype in `subject_annot` from `MethList`.

covariates_string

Columns for covariates to adjust in model from `subject_annot` of `MethList`.

cov.mat

A matrix or list of matrices for random effect in mixed linear model. Default: NULL.

method

`lm` for linear regression, `glm` for generalized linear regression, `mixed` for mixed linear regression. Default: `lm`

family

Which model to fit for generalized and mixed regression models. See glm and fitNullModel

Value

A list containing fitted null model, X and Y matrices, where Y is the phenotypic trait and X created based on `covariates_string`.

Examples

library(tidyverse)
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.4     ✔ readr     2.1.5
#> ✔ forcats   1.0.0     ✔ stringr   1.5.1
#> ✔ ggplot2   3.5.1     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.3     ✔ tidyr     1.3.1
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()
#> ℹ Use the conflicted package (<http://conflicted.r-lib.org/>) to force all conflicts to become errors
library(arrow)
#> 
#> Attaching package: ‘arrow’
#> The following object is masked from ‘package:lubridate’:
#> 
#>     duration
#> The following object is masked from ‘package:utils’:
#> 
#>     timestamp
library(GENESIS)
data(phenoData)
data(chrAnnotation)

wdir <- getwd()
methpath <- system.file('extdata','MethData.csv',package='MethParquet')

# Create Parquet data and MethList
path <- paste0(wdir,'/Parquet_Directory')
write_parquet_meth(data_path=methpath,format='csv',group_by='CHR',parquet_path = path)

mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Fit linear null model (default)
m.null=NullModel(db_obj=mlist,trait='bmi',covariates_string = c('age','sex'))
names(m.null)
#> [1] "NullModel" "Y"         "X"        
m.null$NullModel
#> 
#> Call:
#> lm(formula = y ~ -1 + x_covariates)
#> 
#> Coefficients:
#> x_covariates(Intercept)          x_covariatesage         x_covariatessexM  
#>                44.17322                  0.05353                -11.72508  
#> 
unlink(path,recursive=TRUE)