Create MethList from Amazon Web Services S3 bucket

create_methlist_S3 creates the MethList object linking parquet database stored in Amazon Web Services(AWS) Simple Storage Service (S3) bucket.

create_methlist_S3(
  db_s3_path,
  subject_annot = FALSE,
  subject_path = FALSE,
  cpg_annot = FALSE,
  cpg_path = FALSE,
  subject_readFun = FALSE,
  cpg_readFun = FALSE,
  access_key = FALSE,
  secret_access_key = FALSE,
  default_region = FALSE,
  cpg_col_db,
  subject_col_keep = "all",
  subject_id,
  cpg_col_annot,
  cpgAnnot_col_keep = "all",
  gene_col_name = FALSE,
  chr_col_name = FALSE
)

Arguments

db_s3_path

Filepath to the methylation parquet database in AWS_S3 bucket. Should contain a column named 'CpG' for CpG site names

subject_annot

Subject annotation file

subject_path

Filepath to the subject annotation file in AWS_S3 bucket if not provided as `subject_annot`. Default value: `FALSE`

cpg_annot

CpG annotation file

cpg_path

Filepath to the CpG annotation file in AWS_S3 bucket if not provided as `cpg_annot`. Default value: `FALSE`

subject_readFun

Function to read the subject annotation file in AWS_S3 bucket.

cpg_readFun

Function to read the cpg annotation file in AWS_S3 bucket.

access_key

access key for the AWS_S3 bucket, if applicable.

secret_access_key

secret access key for the AWS_S3 bucket, if applicable.

default_region

AWS default region, if applicable.

cpg_col_db

Column for CpG site names in parquet data, which will be named as `CpG` in CpG data inside meth list

subject_col_keep

Numeric vector containing columns to include in the subject annotation file. Default value: `all`

subject_id

Column for subject IDs in `subject_annot`, which will be named as `sample_id` in `subject_annot` inside meth list

cpg_col_annot

Column for CpG site names in `cpg_annot`

cpgAnnot_col_keep

Numeric vector containing columns to include in the CpG annotation file. Default value: `all`

gene_col_name

Name of column containing gene names in `cpg_annot`. If povided, this column will be named as `Gene` in `cpg_annot` of the meth list. Default value: `FALSE`

chr_col_name

Name of column containing chromosome number in `cpg_annot`. If povided, this column will be named as `CHR` in `cpg_annot` of the meth list. Default value: `FALSE`

Value

A `MethList` object which contains: `db`: open connection to a methylation parquet databse in AWS_S3 `subject_annot`: subject annotation file `cpg_annot`: CpG annotation file

Examples

# Example code only

# mlist=create_methlist_S3(db_s3_path = "path to methylation Parquet data in AWS_S3",
# subject_path="path to subject annotation file in AWS_S3",
# cpg_path="path to CpG annotation file in AWS_S3",
# access_key = "access key to S3 bucket", secret_access_key = "secret access key to S3 bucket",
# default_region = "us-east-1", cpgAnnot_col_keep=c(1:2,12:13,16),
# subject_readFun = load, cpg_readFun = read.csv, cpg_col_db='CpG',
# subject_id='sample_id', cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')