Association analysis with robust linear regression

rlm_ewas_outcome performs association analysis with robust linear regression using methylation as the outcome.

rlm_ewas_outcome(
  db_obj,
  trait,
  select_sites = "full",
  select_chr = FALSE,
  gene_list = FALSE,
  gene_col = FALSE,
  covariates_string,
  parallel = FALSE,
  NAs_to_zero = FALSE,
  out_position = FALSE,
  block_size = 50000
)

Arguments

db_obj

MethList created using create_methlist.

trait

Exposure phenotype in `subject_annot` from `MethList`.

select_sites

Character string consisting of CpG(s) to be tested. Default: `full` (all CpGs). Only one selection for CpG is allowed, `sites`, `chromosome`, or `gene`.

select_chr

Character string of chromosome number to select CpGs from. Default: `FALSE`.

gene_list

Character string of gene names to select CpGs from. Default: `FALSE`.

gene_col

Column for gene names in the CpG annotation data from `Methlist`.

covariates_string

Columns for covariates to adjust in model from `subject_annot` of `MethList`.

parallel

Number of cores used for parallel processing (forking). For details please see plan. Default: FALSE (sequential processes).

NAs_to_zero

Convert missing values to 0 in methylation data. Default: `FALSE`.

out_position

Column name for CpG position in chromosome from `cpg_annot` of `MethList`. If given, output will contain these columns and chromosome number.

block_size

Integer to specify number of CpGs in each iteration block. Default: 50,000.

Value

A data frame with coefficient estimates for `trait` and testing statistics.

Examples

library(tidyverse)
library(arrow)

wdir <- getwd()
methpath <- system.file('extdata','MethData.csv',package='MethParquet')
data(phenoData)
data(chrAnnotation)

# Create Parquet data and MethList
path <- paste0(wdir,'/Parquet_Directory')
write_parquet_meth(data_path=methpath,format='csv',group_by='CHR',parquet_path = path)

mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Run robust linear regression model
ewas_rlm <- rlm_ewas_outcome(db_obj=mlist,trait='sex',select_sites='full',gene_list=FALSE,
select_chr=FALSE,covariates_string = c('bmi','age'),out_position='MAPINFO',NAs_to_zero=TRUE)
head(ewas_rlm)
#>          CpG  F_estimate   p_value    fdr_bh  MAPINFO
#> 1 cg00050692  0.07931863 0.7788291 0.9516462 25524877
#> 2 cg00094412  0.03600158 0.8499128 1.0000000 29592854
#> 3 cg00108715  0.11881806 0.7310740 0.9057686 52565015
#> 4 cg00138407  0.03047686 0.8617803 1.0000000 47386505
#> 5 cg00184953 -0.01347778 1.0000000 1.0000000 31146222
#> 6 cg00222799  0.66834358 0.4156541 0.6731966 43655464
unlink(path,recursive=TRUE)